Fix vars templating on ansible 2.8 and later
On ansible 2.8 and later, vars are not automatically rendered into
hostvars, especially when referencing variables on other inventory
hosts. Playbook vars are also not present in hostvars.

By searching the 'vars' dict instead of 'hostvars' for rules, a
full list of variables can be obtained. From there, templating the
contents of the variable can be done using lookup('vars', varname).

This patch refactors the rule generation to use the vars lookup
plugin to properly template all rules on ansible 2.8 and later.

Closes #8